cleaned up outs

+++ src/delete.py
@@ -2,7 +2,7 @@ import os
 
 states = ['AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NC', 'NE', 'NH', 'NV', 'NJ', 'NM', 'NY', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VA', 'WA', 'WV', 'WI', 'WY', 'global']
 
-dirs = ['jan1','jan2','jan3','jan4','jan5','jan6','feb1','feb2','feb3','feb4','feb5','feb6','mar1','mar2','mar3','mar4','mar5','mar6','apr1','apr2','apr3','apr5','apr6']
+dirs = ['jan1','jan2','jan3','jan4','jan5','feb1','feb4','feb5','mar1','mar2','mar3','mar4','mar5','apr1','apr5']
 
 path = '../tsv/'
 
@@ -10,8 +10,12 @@ def delete():
   for state in states:
     dst = path + state + '/'
     for dir in  dirs:
-      os.remove(dst+dir+'/out.tsv')
-      os.rmdir(dst+dir)
+      f = dst+dir+'/out.tsv'
+      fol = dst+dir
+      if os.path.exists(f):
+        os.remove(dst+dir+'/out.tsv')
+      if os.path.exists(fol):
+        os.rmdir(dst+dir)
 
 if __name__ == '__main__':
   delete()
